setbatterypercent now takes the place of battery drainage commands
A more forall-oriented approach
